Backport 752bc1f2 to fix dev-dep stripping
The releng branch already has this fix as well as a bunch of other
changes. However substrate still uses the master branch and swithing it
is a non trivial task due to breaking chnanges.

So as a quick solution, backport this fix while we work on switching
every one over to the new branch.

Fixes #37